+++ Makefile
@@ -16,13 +16,13 @@ crate-tests: crate-bindings-binaries crate-bindings-build
 .PHONY: crate-bindings-binaries
 crate-bindings-binaries: export FORCE_SKIA_BINARIES_DOWNLOAD=1
 crate-bindings-binaries:
-	cd skia-bindings && cargo publish -vv --dry-run --features "gl,vulkan,textlayout"
+	cd skia-bindings && cargo publish -vv --dry-run --features "gl,vulkan,textlayout,binary-cache"
 	cd skia-bindings && cargo publish -vv --dry-run 
 
 .PHONY: crate-bindings-build
 crate-bindings-build: export FORCE_SKIA_BUILD=1
 crate-bindings-build: 
-	cd skia-bindings && cargo publish -vv --dry-run --features "gl,vulkan,textlayout,d3d"
+	cd skia-bindings && cargo publish -vv --dry-run --no-default-features --features "gl,vulkan,textlayout,d3d"
 	cd skia-bindings && cargo publish -vv --dry-run 
 
 .PHONY: publish
