+++ skia-bindings/build_support.rs
@@ -1,14 +1,13 @@
 //! Build support functions for the Rust-Skia library.
 
 pub mod android;
-pub mod azure;
-pub mod binaries;
 pub mod cargo;
 pub mod clang;
-pub mod git;
 pub mod ios;
 pub mod llvm;
 pub mod skia;
-pub mod utils;
 pub mod vs;
 pub mod xcode;
+
+#[cfg(feature = "binary-cache")]
+pub mod binary_cache;
