+++ skia-safe/Cargo.toml
@@ -20,7 +20,7 @@ edition = "2018"
 doctest = false
 
 [features]
-default = []
+default = ["binary-cache"]
 gl = ["gpu", "skia-bindings/gl"]
 egl = ["gl", "skia-bindings/egl"]
 x11 = ["gl", "skia-bindings/x11"]
@@ -32,6 +32,8 @@ textlayout = ["skia-bindings/textlayout"]
 webp = ["webp-encode", "webp-decode"]
 webp-encode = ["skia-bindings/webp-encode"]
 webp-decode = ["skia-bindings/webp-decode"]
+use-system-jpeg-turbo = ["skia-bindings/use-system-jpeg-turbo"]
+binary-cache = ["skia-bindings/binary-cache"]
 
 # implied only, do not use
 gpu = []
@@ -42,7 +44,7 @@ shaper = ["textlayout", "skia-bindings/shaper"]
 [dependencies]
 bitflags = "1.2"
 lazy_static = "1.4"
-skia-bindings = { version = "=0.40.0", path = "../skia-bindings" }
+skia-bindings = { version = "=0.40.0", path = "../skia-bindings", default-features = false }
 # for d3d types
 winapi = { version = "0.3.9", features = ["d3d12", "dxgi"], optional = true }
 # for ComPtr
